requirements.txt
httpx==0.27.0
PyYAML==6.0.1
APScheduler==3.10.4
pandas==2.2.2
pandas-ta==0.3.14b
